API Documentation

API Documentation

General Utilities

List of the General Functions and Types provided by the RayTracer. Most of the other functionalities are built upon these.

RayTracer.Vec3 — Type.

This is the central type for RayTracer. All of the other types are defined building upon this.

All the fields of the Vec3 instance contains Arrays. This ensures that we can collect the gradients w.r.t the fields using the Params API of Zygote.

Defined Operations for Vec3:

+, -, * – These operations will be broadcasted even though there is no explicit mention of broadcasting.

dot, l2norm

cross

clamp, clip01

zero, similar, one

place

maximum, minimum

source

RayTracer.rgb — Type.

rgb is an alias for Vec3. It makes more sense to use this while defining colors.

source

RayTracer.get_image — Method.

get_image(image, width, height)

Reshapes and normalizes a Vec3 color format to the RGB format of Images for easy loading, saving and visualization.

source

RayTracer.zeroonenorm — Method.

zeroonenorm(x::AbstractArray)

Normalizes the elements of the array to values between 0 and 1. It is recommended to use ths function for doing this simple operation because the adjoint for this function is incorrect in Zygote.

source

RayTracer.FixedParams — Type.

FixedParams

Any subtype of FixedParams is not optimized using the update! API. For example, we don't want the screen size to be altered while inverse rendering, this is ensured by wrapping those parameters in a subtype of FixedParams.

source

RayTracer.@diffops — Macro.

@diffops MyType::DataType

Generates functions for performing gradient based optimizations on this custom type. 5 functions are generated.

x::MyType + y::MyType – For Gradient Accumulation

x::MyType - y::MyType – For Gradient Based Updates

x::MyType * η<:Real – For Multiplication of the Learning Rate with Gradient

η<:Real * x::MyType – For Multiplication of the Learning Rate with Gradient

x::MyType * y::MyType – Just for the sake of completeness.

Most of these functions do not make semantic sense. For example, adding 2 PointLight instances do not make sense but in case both of them are gradients, it makes perfect sense to accumulate them in a third PointLight instance.

source

RayTracer.bigmul — Method.

bigmul(x)

Returns the output same as typemax. However, in case gradients are computed, it will return the gradient to be 0 instead of nothing as in case of typemax.

source

RayTracer.extract — Method.

extract(cond, x<:Number)
extract(cond, x<:AbstractArray)
extract(cond, x::Vec3)

Extracts the elements of x (in case it is an array) for which the indices corresponding to the cond are true.

Note

extract has a performance penalty when used on GPUs.

Example:

julia> a = rand(4)
4-element Array{Float64,1}:
 0.7201598586590607 
 0.5829718552672327 
 0.1177531256556108 
 0.3083157590071375 

julia> cond = a .> 0.5
4-element BitArray{1}:
  true
  true
 false
 false

julia> RayTracer.extract(cond, a)
2-element Array{Float64,1}:
 0.7201598586590607
 0.5829718552672327

source

RayTracer.improcess — Method.

improcess(color_vector, width, height)

Normalizes the Color Dimension and reshapes it to the given configuration.

source

RayTracer.place — Method.

place(a::Vec3, cond)

Constructs a new Vec3 with array length equal to that of cond filled with zeros. Then it fills the positions corresponding to the true values of cond with the values in a.

The length of each array in a must be equal to the number of true values in the cond array.

source

RayTracer.get_params — Function.

get_params(x)

Get the parameters from a struct that can be tuned. The output is in the form of an array.

source

RayTracer.set_params! — Function.

set_params!(x, y::AbstractArray)

Sets the tunable parameters of the struct x. The index of the last element set into the struct is returned as output. This may be used to confirm that the size of the input array was as expected.

Example

julia> scene = Triangle(Vec3(-1.9, 1.3, 0.1), Vec3(1.2, 1.1, 0.3), Vec3(0.8, -1.2, -0.15),
                    color = rgb(1.0, 1.0, 1.0), reflection = 0.5)
Triangle{Array{Float64,1}}(Vec3{Array{Float64,1}}([-1.9], [1.3], [0.1]), Vec3{Array{Float64,1}}([1.2], [1.1], [0.3]), Vec3{Array{Float64,1}}([0.8], [-1.2], [-0.15]), RayTracer.Material{RayTracer.PlainColor,Float64}(RayTracer.PlainColor(Vec3{Array{Float64,1}}([1.0], [1.0], [1.0])), 0.5))

julia> x = rand(13)
13-element Array{Float64,1}:
 0.39019817669623835
 0.940810689314205
 .
 .
 .
 0.5590307650917048
 0.7551647340674075

julia> RayTracer.set_params!(scene, x)
13

source

Differentiation

The recommended mode of differentation is by Automatic Differentiation using Zygote. Refer to the Zygote docs for this. The API listed below is for numerical differentiation and is very restrictive in its current form.

RayTracer.ngradient — Function.

ngradient(f, xs::AbstractArray...)

Computes the numerical gradients of f w.r.t xs. The function f must return a scalar value for this to work. This function is not meant for general usage as the value of the parameter δ has been tuned for this package specifically.

Also, it should be noted that these gradients are highly unstable and should be used only for confirming the values obtained through other methods. For meaningful results be sure to use Float64 as Float32 is too numerically unstable.

source

RayTracer.numderiv — Function.

numderiv(f, θ)
numderiv(f, θ::AbstractArray)
numderiv(f, θ::Real)

Compute the numerical derivates wrt one of the scene parameters. The parameter passed cannot be enclosed in an Array, thus making this not a general method for differentiation.

Note

This is not a generalized method for getting the gradients. For that please use Zygote. However, this can be used to debug you model, incase you feel the gradients obtained by other methods is sketchy.

source

Scene Configuration

Camera

Light

Materials

Objects

Renderers

RayTracer.raytrace — Method.

raytrace(origin::Vec3, direction::Vec3, scene::Vector, lgt::Light, eye_pos::Vec3, bounce::Int)
raytrace(origin::Vec3, direction::Vec3, scene::Vector, lgt::Vector{Light}, eye_pos::Vec3, bounce::Int)

Computes the color contribution to every pixel by tracing every single ray. Internally it calls the light function which implements Blinn Phong Rendering and adds up the color contribution for each object.

The eye_pos is simply the origin when called by the user. However, the origin keeps changing across the recursive calls to this function and hence it is necessary to keep track of the eye_pos separately.

The bounce parameter allows the configuration of global illumination. To turn off global illumination set the bounce parameter to >= 2. As expected rendering is much faster if global illumination is off but at the same time is much less photorealistic.

Note

The support for multiple lights is primitive as we loop over the lights. Even though it is done in a parallel fashion, it is not the best way to do so. Nevertheless it exists just for the sake of experimentation.

source

RayTracer.light — Method.

light(s::Object, origin, direction, dist, lgt::Light, eye_pos, scene, obj_num, bounce)

Implements the Blinn Phong rendering algorithm. This function is merely for internal usage and should in no case be called by the user. This function is quite general and supports user defined Objects. For support of custom Objects have a look at the examples.

source

Optimization

RayTracer.update! — Method.

update!(opt, x, Δ)

Provides an interface to use all of the Optimizers provided by Flux. The type of x can be anything as long as the operations defined by @diffops are available for it. By default all the differentiable types inside the Package can be used with it.

The type of Δ must be same as that of x. This prevent silent type conversion of x which can significantly slow doen the raytracer.

Example:

opt = ADAM()

gs = gradient(loss_function, θ)

update!(opt, θ, gs[1])

source