High-level API

Raytracer.demo — Method

demo(;output_file::String = "demo.jpg",
      camera_type::Type{<:Camera} = PerspectiveCamera,
      camera_position::Tuple{Float32, Float32, Float32} = (-3f0, 0f0, 0f0),
      camera_orientation::Tuple{Float32, Float32, Float32} = (0f0, 0f0, 0f0),
      screen_distance::Float32 = 2f0,
      image_resolution::Tuple{Int, Int} = (540,540),
      samples_per_side::Int = 0,
      renderer_type::Type{<:Renderer} = PathTracer,
      pt_n::Int = 10,
      pt_max_depth::Int = 2,
      pt_roulette_depth::Int = 3,
      α::Float32 = 0.75f0,
      γ::Float32 = 1f0,
      use_threads::Bool = true,
      disable_output::Bool = false)

Create a demo image.

Arguments

output_file: output LDR file name (the HDR file will have the same name, but with "pfm" extension).

camera_type: choose camera type ("perspective" or "orthogonal").

camera_position: camera position in the scene as "X,Y,Z".

camera_orientation: camera orientation as "angX,angY,angZ".

screen_distance: only for "perspective" camera: distance between camera and screen.

image_resolution: resolution of the rendered image.

renderer_type: type of renderer to use ("onoff", "flat" or "path").

samples_per_side: number of samples per pixel (must be a perfect square).

pt_n: number of rays fired for mc integration (path-tracer only).

pt_max_depth: maximum number of reflections for each ray (path-tracer only).

pt_roulette_depth: depth of the russian-roulette algorithm (path-tracer only).

α: scaling factor for the normalization process.

γ: gamma value for the tone mapping process.

use_threads: use macro Threads.@threads.

disable_output: if true, no message is printed.

Raytracer.tonemapping — Method

tonemapping(input_file::String,
            output_file::String,
            α::Float32,
            γ::Float32
            ; disable_output::Bool = false)

Load a pfm hdr image, apply the tone mapping, and save the generated ldr image.

input_file is the path to a valid PFM image. output_file is the path in which save the generated image + the name of the output file. The output format is deduced by the extension of the output file: if the name is example.jpg, the image will be exported in JPEG format. α is the normalization coefficient. γ is the actor of the γ correction.

If disable_output is true, no message is printed.

Index

Raytracer.demo

Raytracer.tonemapping